+++ src/mock/fault.rs
@@ -60,6 +60,11 @@ const ONVIF_ERROR: &str = "http://www.onvif.org/ver10/error";
 pub(super) const INVALID_ARG_VAL: FaultQName = FaultQName::new("ter", ONVIF_ERROR, "InvalidArgVal");
 pub(super) const ACTION: FaultQName = FaultQName::new("ter", ONVIF_ERROR, "Action");
 pub(super) const NO_PROFILE: FaultQName = FaultQName::new("ter", ONVIF_ERROR, "NoProfile");
+pub(super) const NO_CONFIG: FaultQName = FaultQName::new("ter", ONVIF_ERROR, "NoConfig");
+pub(super) const PROFILE_EXISTS: FaultQName = FaultQName::new("ter", ONVIF_ERROR, "ProfileExists");
+pub(super) const MAX_PROFILES: FaultQName = FaultQName::new("ter", ONVIF_ERROR, "MaxNVTProfiles");
+pub(super) const CONFIGURATION_CONFLICT: FaultQName =
+    FaultQName::new("ter", ONVIF_ERROR, "ConfigurationConflict");
 pub(super) const DELETION_OF_FIXED_PROFILE: FaultQName =
     FaultQName::new("ter", ONVIF_ERROR, "DeletionOfFixedProfile");
 pub(super) const WELL_FORMED: FaultQName = FaultQName::new("ter", ONVIF_ERROR, "WellFormed");